King county housing data
The objective of this project is to understand what factors impact the sales prices of house. Data is limited to only 2014-2015. It doesn't consider the recent changes of housing market and demand.  For purupose of EDA. log of price values is considered. Reviewed plots and identified the factors and developed models to predict the sales price.